pc_app/requirements.txt
# VR Screen Streamer - PC Application Dependencies
# Install with: pip install -r requirements.txt

# GUI Framework
PyQt5>=5.15.0

# Screen Capture
mss>=9.0.0
Pillow>=10.0.0

# Video Encoding/Compression
numpy>=1.24.0
opencv-python>=4.8.0

# WebSocket Server for streaming
websockets>=12.0
aiohttp>=3.9.0

# Network utilities
netifaces>=0.11.0

# Windows-specific for window enumeration
pywin32>=306

# Performance monitoring
psutil>=5.9.0

# Configuration
pyyaml>=6.0.0
